Finished the TimeUntilAlarm functionality.
Took some time, but it was a great learning process. I had to google a lot of questions on Stack Overflow since I don't have that much experience with Swift.

Overall very fun to see the final result work. Next part is to actually be able to go back to the screen and add the selected alarm time to the dynamic table view cell.