Add Data to Alarm Table View Success
I've successfully implemented adding data from another class to my main alarm table view. So when a user presses the + button, he/she can specify when the alarm will go off. Then from the previous commit, the time until alarm will automatically calculate based on the most recent time chosen.

Those values are then used to create an Alarm instance, only when the user clicks done. After that, in the main Alarm Table View class, it will append the instance of the new alarm to the array (the data). Then since we are dealing with UITableView, we need to insert a row with the new alarm details. The function is: insertRows(at: [IndexPath], with: UITableViewRowAnimation). The key to understanding this, is the IndexPath, so before calling insertRows we need to create a IndexPath, which is basically a NSIndexPath (Obj-c). So we create an index path like this: let indexPath = NSIndexPath(forRow: alarms.count-1, inSection:0).

Important part part 2 lol, the forRow here is referring to the last element in the array, in other words the alarm instance we have just added to the array. That is how it inserts the alarm details in the new row.

Next step: Is to add a text field so that the user can name the alarm as they wish. Also delete the sample data so the user starts fresh.