TextField - Alarm Name Success
I've added a text field so that users can enter an alarm name of their choice and once they click Done, it will save that name. I've played around with the delegate of UITextField to close the keyboard when the user either presses "enter" or "done". I've also deleted the second section which had nothing in it.

Next Step: to close keyboard when the user taps anywhere on the current screen.