Switch Toggled - Change Cell Color Success
On this commit, I allowed users to tap anywhere in the Table View to exit the keyboard when entering an Alarm name. I have used UIGestureRecognizer to recognize when a user taps on the Table View and then calling the resignFirstResponder() method on the textField. That part was very quick to implement.

Then my next implementation was to change the color of the Table Cell depending if the switch (of the Alarm) was on or off. If the switch is on, the cell color should be Green, if switch is off, White. Although in the future I would like to have my main interface be all black instead of white. Possibly even let the user switch between themes. This implementation took some time since I was confused on how to make the switch Action and cell Color interact with each other. After searching through searches of people with the same problem, I found a Stack Overflow post that explained very well how to use and implement NotificationCenter. It turns out its very easy to interact between classes with NotificationCenter. I will have to look at the docs for this, since I've just used it briefly without really knowing what I am doing.

Next step: To allow user to click on any cell and edit that cell's Alarm data.