Fixed a lot of bugs!
What I did: Editing Cell works! Cancel button works! No delay in Alarm Time (First View)

Details: Since my last commit, I've changed quite a bit of stuff. I've encontered a lot more bugs than all the other commits. One of the obstacles I encountered was manipulating my data array from multiple view controllers. Since I did not know the big idea behind the process I had to go through a lot of obstacles to finally understanding. I had my array as a separate class (from following a tutorial), although that might prove useful for bigger applications, it was quite unnecessary for my app.

I just initialized my Alarms Array in:  AlarmTableViewController. This makes sense since this is the table with all the alarms. When I either clicked on the cell or the plus button, in the prepare for segue function, I set the other class (MakeAlarmTableViewController) alarms array to the current array. This will allow me to set the Date, Alarm Name and Time Until Alarm correctly from what is in the array, more specifically at the given index (index depends on which cell is selected).

Finally (this is like an essay haha), I changed the ViewController timer to be fired off at each 0.01 seconds so that there is no delay when displaying the Time, again this is for the first screen that the user sees.

Next Step: I would like less delay when one clicks any cell to edit it. I've tested it and it seems that the time delay is always different on each click. Safe to say, it is not yet super user friendly. Also next step would be to continuously re-arrange cells based on the correct Time order (from AM to PM).