Night Color - Success
I have changed the interface UI. Instead of the boring white color, I've changed everything to revolve around darker colours. Much easier on the eye at night. Of course this is only the beginning of adding themes and letting the user choose. However I've had difficulties with setting the navigation bar black and the status bar text to white. After much research I've found a way. Also there was a big bug in which if the user decided to go back to the main screen, when swiping up to the alarm table view, the array would just reset itself to an empty array. To fix this, I've added the array in the SampleData class and created a getter function. Whenever the alarm table view appears, it gets the array from Sample Array.

Next step: To figure out the best method to implement themes. I will most likely need a class dedicated to themes itself. To figure out how to make the alarm work is also the next step.