Theme Works in all Views!
I have modified the appearance of all views based on which theme is selected. Also a big change added is the custom switch on the initial view. It has a very nifty design (not mine) and the animation is very nice. The user can clearly make the distinction between night and day option on the switch. All views will adapt to the user's selection.

Next step: To read on NSUserDefaults because currently my array resets every time i reset my app and the switch state also resets itself.