User Choose Song Success!
After a good week or two of going at this implementation, I've successfully implemented the ability the user to choose a song from their iPhone music library or a default song provided by the app. A lot of new features in this commit.

Main features:

•When user choses a song, appropriate data is stored in respective Alarm array to later play the song.
•Added a new Table View Controller to allow user to choose a song.
•Time Until Alarm Label is continually updated (each minute)

Details:

The selected song is saved into the respective Alarm array. If user chooses a song from iPhone music library, a MPMediaItemCollection is stored in the array. If user chooses a default song, a URL is stored the array. Getting the string title of the song for the MPMediaItemCollection was easy, each object stores a variable title.

However getting the title for the URL was more difficult since URL's don't have a static title variable. In my AddSongTableViewController I had created an array that contained all my cells. I represented my cells as structs, each had a type and title. Since the defaults songs were static I stored all titles. To get the default song title I decided to store the default song as a dictionary which contained the title and URL. Only one song is chosen, so the dictionary will always be of length 1. Hence for other classes getting the key of the dictionary, order wouldn't be a problem.

I also worked with xib files to create two custom cells (the cells that contained headers).

Overall: This was definitely the most exhausting sprint I've done. I'm glad I've finally reached a point where I can compile and choose a song without any bugs.

Next Steps: Allow user to play the song chosen. Fix layout constraints on custom xib cells.