Song Selected Checkmark Success!
When a user selects a song, a checkmark appears beside the selected song. This allows for more clarity on which song the user has chosen. I've had to clean a lot of code and go through my table view controller files several times. Passing data between both (Alarm & Add Song) Controllers, I was able to put a checkmark on the current song and update that checkmark accordingly. The difficulty in this commit was trying to handle all possible cases. There are a lot of factors deciding where to put the checkmark. Example, did the user click on a edited cell, or did he create a new cell but already chose a song and is looking to change that song.

Also an important change was that the songTitleData (the custom songs list from the user's iPhone library) is now behaving like a stack (FIFO). When a user chooses a new song and the list is full, the first song disappears and the new song is appended to the end. This is the same behaviour as in the native iPhone Clock app.

Next Steps: When user clicks "back" to go to Alarm Controller, don't save the current settings, revert back to the old Alarm. Also update UI.